Fix bad URLs sent out in welcome email
Welcome emails were sent with the :adventure_id param set to the DB ID (a UUID), not the global ID. Whoops.

This works around the problem for existing emails by checking the incoming "id" value and converting it to a relay Global ID if it looks like a UUID.

This also fixes the issue for future emails by generating the global ID when building the URL in the email template.